# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pakistani Job Aggregator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user's profile
      responses:
        "200":
          description: Profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profile/cv:
    post:
      operationId: uploadCv
      tags: [profile]
      summary: Upload CV (PDF or DOCX) — extracts text and generates embedding
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CvFileInput"
      responses:
        "200":
          description: CV processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CvUploadResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /saved-jobs:
    get:
      operationId: listSavedJobs
      tags: [savedJobs]
      summary: List all saved/tracked jobs for the current user
      responses:
        "200":
          description: Saved jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedJob"
    post:
      operationId: saveJob
      tags: [savedJobs]
      summary: Save a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedJobInput"
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedJob"
  /saved-jobs/{id}:
    patch:
      operationId: updateSavedJob
      tags: [savedJobs]
      summary: Update status or notes on a saved job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedJobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedJob"
    delete:
      operationId: deleteSavedJob
      tags: [savedJobs]
      summary: Remove a saved job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
    CvFileInput:
      type: object
      properties:
        file:
          type: string
          format: binary
      required:
        - file
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    UserPreferences:
      type: object
      properties:
        sectors:
          type: array
          items:
            type: string
        locations:
          type: array
          items:
            type: string
        experienceLevel:
          type: string
        minMatchScore:
          type: number
    Profile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        cvText:
          type: ["string", "null"]
        cvUpdatedAt:
          type: ["string", "null"]
        preferences:
          $ref: "#/components/schemas/UserPreferences"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
    ProfileUpdate:
      type: object
      properties:
        preferences:
          $ref: "#/components/schemas/UserPreferences"
    CvUploadResult:
      type: object
      properties:
        success:
          type: boolean
        cvTextLength:
          type: integer
        cvUpdatedAt:
          type: string
      required:
        - success
        - cvTextLength
        - cvUpdatedAt
...
    Job:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        title:
          type: string
        company:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        sector:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        requirements:
          type: ["string", "null"]
        experienceLevel:
          type: ["string", "null"]
        jobType:
          type: ["string", "null"]
        salaryRange:
          type: ["string", "null"]
        postedDate:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        applyUrl:
          type: ["string", "null"]
        applicationSteps:
          type: array
          items:
            type: string
        matchScore:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - source
        - title
        - createdAt
...
    SavedJob:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        jobId:
          type: string
        job:
          $ref: "#/components/schemas/Job"
        status:
          type: string
          enum: [saved, applied, interview, rejected, offer]
        notes:
          type: ["string", "null"]
        appliedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - jobId
        - status
        - createdAt
    SavedJobInput:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [saved, applied, interview, rejected, offer]
        notes:
          type: string
      required:
        - jobId
    SavedJobUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [saved, applied, interview, rejected, offer]
        notes:
          type: string
        appliedAt:
          type: string
...