# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pakistani Job Aggregator API
 * OpenAPI spec version: 0.1.0
 */
import type { UserPreferences } from './userPreferences';

export interface Profile {
  id: string;
  userId: string;
  /** @nullable */
  cvText?: string | null;
  /** @nullable */
  cvUpdatedAt?: string | null;
  preferences?: UserPreferences;
  createdAt?: string;
  updatedAt?: string;
}