# lib/api-zod/src/generated/types/savedJob.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pakistani Job Aggregator API
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from './job';
import type { SavedJobStatus } from './savedJobStatus';

export interface SavedJob {
  id: string;
  userId: string;
  jobId: string;
  job?: Job;
  status: SavedJobStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  appliedAt?: string | null;
  createdAt: string;
}