# lib/api-zod/src/generated/types/savedJobInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pakistani Job Aggregator API
 * OpenAPI spec version: 0.1.0
 */
import type { SavedJobInputStatus } from './savedJobInputStatus';

export interface SavedJobInput {
  jobId: string;
  status?: SavedJobInputStatus;
  notes?: string;
}